requirements.txt
confluent-kafka==2.3.0
SQLAlchemy==2.0.25
pymysql==1.1.0
pydantic==2.6.1
joblib==1.3.2
python-dotenv==1.0.1
scikit-learn==1.3.2
cryptography
pytest==8.0.0
streamlit==1.32.0
